食べもの図鑑の表示 食べ物、食べた物登録ボタン追加

+++ app/src/main/java/com/gmail/gogoty913/macromanagetool/entity/FoodInfo.java
@@ -38,9 +38,9 @@ public class FoodInfo {
     public double carbohydrate;
 
     @ColumnInfo
-    public int displayCapacity;
+    public double displayCapacity;
 
     @ColumnInfo
-    public int allCapacity;
+    public double allCapacity;
 
 }
